requirements.txt
google 
python-dotenv 
langchain
langchain_google_genai
google.generativeai
google-cloud-vision
google-cloud-texttospeech
speechrecognition
pillow
sounddevice
soundfile
numpy
streamlit
requests
typing-extensions
langchain
langchain-core
google-generative-ai
pydantic
